worker/requirements.txt
amqp==5.0.9
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Flask==2.0.3
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
logging